\name{BaylorEdPsych-package}
\alias{BaylorEdPsych-package}
%\alias{BaylorEdPsych}
\docType{package}
\title{
R functions/data for Baylor Ed Psych Department's Quantitative Courses
}
\description{
R functions and data for Baylor University Educational Psychology Department's Quantitative Courses
}
\details{
\tabular{ll}{
Package: \tab BaylorEdPsych\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-08\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
A. Alexander Beaujean

Maintainer: A. Alexander Beaujean <Alex_Beaujean@Baylor.edu>
}



