\name{BaylorEdPsych}
\alias{BaylorEdPsych}
\title{Baylor Educational Psychology Quantitative Courses R Functions and Data}


\description{Baylor Educational Psychology Quantitative Courses R Functions and Data}
	
\value{	
Data Sets:
\item{\code{\link{MLBOffense2011}}}{2011 Major League Baseball Offensive Data}	
\item{\code{\link{MLBPitching2011}}}{2011 Major League Baseball Pitching Data}	
\item{\code{\link{EndersTable1_1}}}{Data from table 1.1 (p. 3) in Enders, C. K. (2011). \emph{Applied missing data analysis}. New York: Guilford}


Effect Sizes:
\item{\code{\link{PseudoR2}}}{Various Pseudo-\eqn{R^2} values for a regression with a dichotomous outcome}
\item{\code{\link{CohensD.rawData}}}{Cohen's d effect size from raw data}
\item{\code{\link{CohensD.unpairedT}}}{Cohen's d effect size from unpaired (independent samples) t test}
\item{\code{\link{HedgesG.rawData}}}{Hedge's g effect size from raw data}

Missing Data:
\item{\link{LittleMCAR}}{Little's Missing Completely At Random (MCAR) chi-square test}	

Data Preperation:
\item{\link{BilogData}}{Coverts a R data frame to a fixed width data frame for use in BILOG-MG and exports it to a \code{.dat} file}	

}


\references{
Baylor Psychometric Laboratory: \url{http://blogs.baylor.edu/psychometric_lab/}

Baylor University Educational Psychology: \url{http://www.baylor.edu/soe/edp/}
}

\author{
A. Alexander Beaujean
}


\keyword{ Missing Data }
\keyword{ Effect Sizes }
\keyword{ Pseudo-R2 }