\name{EndersTable1_1}
\alias{EndersTable1_1}
\docType{data}
\title{
Data from Table 1.1 (p. 3) of Enders (2011) \emph{Applied Missing Data Analysis}
}
\description{
Data from Table 1.1 (p. 3) of Enders (2011) \emph{Applied Missing Data Analysis}.
}
\usage{data(EndersTable1_1)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{IQ}}{IQ}
    \item{\code{JP}}{Job Performace}
    \item{\code{WB}}{Psychological Well Being}
  }
}

\source{
Enders, C. K. (2011). \emph{Applied missing data analysis}. New York: Guilford.
}

\examples{
data(EndersTable1_1)
}

\keyword{datasets}
\keyword{Missing Data}
